Move cluster state to separate package
