+++ web/src/collection/state/sagas.js
@@ -1,70 +1,15 @@
-import { fork, call, put, select, takeLatest } from "redux-saga/effects";
+import { call, fork, put, select, takeLatest } from "redux-saga/effects";
 import {
-  ACTION_FETCH_FILE_CLUSTER,
   ACTION_FETCH_FILES,
-  ACTION_UPDATE_FILE_CLUSTER_FILTERS,
   ACTION_UPDATE_FILTERS,
-  fetchFileClusterFailure,
-  fetchFileClusterSuccess,
   fetchFilesFailure,
   fetchFilesSuccess,
-  updateFileClusterFiltersFailure,
-  updateFileClusterFiltersSuccess,
   updateFiltersFailure,
   updateFiltersSuccess,
 } from "./actions";
 import { selectColl, selectFileCluster, selectFileMatches } from "./selectors";
 import fileMatchRootSaga from "./fileMatches/sagas";
-
-function* updateFileClusterFiltersSaga(server, action) {
-  yield* fetchFileClusterSaga(
-    server,
-    action,
-    updateFileClusterFiltersSuccess,
-    updateFileClusterFiltersFailure
-  );
-}
-
-function* fetchFileClusterPageSaga(server, action) {
-  yield* fetchFileClusterSaga(
-    server,
-    action,
-    fetchFileClusterSuccess,
-    fetchFileClusterFailure
-  );
-}
-
-function* fetchFileClusterSaga(server, action, success, failure) {
-  try {
-    // Determine current query params
-    const { limit, filters, matches: current } = yield select(
-      selectFileCluster
-    );
-
-    // Send request to the server
-    const resp = yield call([server, server.fetchFileCluster], {
-      limit,
-      offset: current.length,
-      id: filters.fileId,
-      fields: filters.fields,
-      filters,
-    });
-
-    // Handle error
-    if (resp.failure) {
-      console.error("Fetch file matches error", resp.error);
-      yield put(failure(resp.error));
-      return;
-    }
-
-    // Update state
-    const { total, matches, files } = resp.data;
-    yield put(success(matches, files, total));
-  } catch (error) {
-    console.error(error);
-    yield put(failure(error));
-  }
-}
+import fileClusterRootSaga from "./fileCluster/sagas";
 
 function resolveReportActions(fetchAction) {
   switch (fetchAction.type) {
@@ -117,16 +62,11 @@ function* fetchFilesSaga(server, action) {
  */
 export default function* collRootSaga(server) {
   yield fork(fileMatchRootSaga, server, selectFileMatches);
+  yield fork(fileClusterRootSaga, server, selectFileCluster);
 
   yield takeLatest(
     [ACTION_UPDATE_FILTERS, ACTION_FETCH_FILES],
     fetchFilesSaga,
     server
   );
-  yield takeLatest(
-    ACTION_UPDATE_FILE_CLUSTER_FILTERS,
-    updateFileClusterFiltersSaga,
-    server
-  );
-  yield takeLatest(ACTION_FETCH_FILE_CLUSTER, fetchFileClusterPageSaga, server);
 }
